Add responsiveness to active users section
You can now scroll to see more users in smaller displays. I also
changed the z-index of the green active status circles so that they
would not show up through the header. Did a similar thing to the
floating message circles so that they had a higher index than the
content in the feed.